pyproject.toml
[tool.poetry]
name = "astroai"
version = "0.1.0"
description = "Astro Conversational Insight Agent"
authors = ["AI <ai@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
fastapi = ">=0.111.0"
uvicorn = {version = ">=0.30.0", extras = ["standard"]}
pydantic = ">=2.7.0"
langchain = ">=0.2.0"
langchain-core = ">=0.2.0"
langchain-openai = ">=0.1.0"
langchain-community = ">=0.2.0"
langchain-chroma = ">=0.1.0"
langgraph = ">=0.0.60"
redis = ">=5.0.0"
chromadb = ">=0.5.0"
python-dotenv = ">=1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
